DQ dashboard data add prev statistics

+++ dashboard/export_dq_marts.py
@@ -1,4 +1,5 @@
 import os
+import sys
 from utils import yaml
 from pathlib import Path
 import snowflake.connector as sf
@@ -12,7 +13,6 @@ DBT_PROFILE_NAME = "dq_demo"
 profile_config = None
 with open(f"{DBT_PROFILE_DIR}/profiles.yml", 'r') as file:
     profile_config = yaml.load_yaml_text(file.read())
-print(profile_config)
 profile_config = profile_config.get(DBT_PROFILE_NAME, {})
 target = profile_config.get("target")
 config = profile_config.get("outputs", {}).get(target, {})
@@ -24,16 +24,20 @@ for key, value in config.items():
 # Export table configured at args[1]
 with sf.connect(**config) as conn:
     data = pd.read_sql(sql=f"select * from {config.get('database')}.{config.get('schema')}.BI_DQ_METRICS", con=conn)
-print("--data:", data)
+print("--data:", data.shape)
 
 # Write to Duck DB
 db_file = f"./dashboard/dq_mart.duckdb"
+incremental = eval(sys.argv[1]) if len(sys.argv) > 1 else True
+assert isinstance(incremental, bool)
+print("--incremental:", incremental)
 with duckdb.connect(db_file) as duckdb_conn:
     existed = not duckdb_conn\
         .sql(f"select * from information_schema.tables where table_name='BI_DQ_METRICS'")\
         .to_df().empty
     if existed:
-        duckdb_conn.sql(f"truncate table BI_DQ_METRICS")
+        if not incremental:
+            duckdb_conn.sql(f"truncate table BI_DQ_METRICS")
     else:
         duckdb_conn.sql(f"create table BI_DQ_METRICS as select * from data")
     duckdb_conn.sql(f"insert into BI_DQ_METRICS select * from data")
